Split street/number flipping and language translation into separate modules
