+++ middleware/localNamingConventions.js
@@ -1,10 +1,7 @@
-
 var check = require('check-types');
 var _ = require('lodash');
-var logger = require('pelias-logger').get('api:middleware:localNamingConventions');
 
 var flipNumberAndStreetCountries = ['DEU', 'FIN', 'SWE', 'NOR', 'DNK', 'ISL', 'CZE'];
-var translations = {};
 
 function setup() {
   var api = require('pelias-config').generate().api;
@@ -13,9 +10,6 @@ function setup() {
     if (localization.flipNumberAndStreetCountries) {
       var countries = localization.flipNumberAndStreetCountries;
       flipNumberAndStreetCountries = _.uniq(flipNumberAndStreetCountries.concat(countries));
-    }
-    if (localization.translations) {
-      translations = require(localization.translations);
     }
   }
   return applyLocalNamingConventions;
@@ -46,38 +40,7 @@ function applyLocalNamingConventions(req, res, next) {
   })
   .forEach( flipNumberAndStreet );
 
-  var lang;
-  if (req.clean) {
-    lang = req.clean.lang;
-  }
-
-  if( lang && translations[lang] ) {
-    _.forEach(translations[lang], function(names, key) {
-      _.forEach(res.data, function(place) {
-        translateName(place, key, names);
-        translateName(place.parent, key, names);
-      });
-    });
-  }
   next();
-}
-
-
-function translateName(place, key, names) {
-  if( place[key] !== null ) {
-    var name;
-    if (place[key] instanceof Array) {
-      name = place[key][0];
-      if (name && names[name]) {
-        place[key][0] = names[name]; // do the translation
-      }
-    } else {
-      name = place[key];
-      if (name && names[name]) {
-        place[key] = names[name];
-      }
-    }
-  }
 }
